+++ tests/clients/test_alertrules.py
@@ -10,6 +10,7 @@ from py42.services.alertrules import AlertRulesService
 from py42.services.alerts import AlertService
 
 TEST_RULE_ID = "rule-id"
+TEST_USER_ID = "user-id"
 
 
 TEST_SYSTEM_RULE_RESPONSE = {
@@ -43,21 +44,18 @@ def mock_alerts_service(mocker):
 
 
 class TestAlertRulesClient(object):
-    _rule_id = u"test-rule-id"
-    _user_id = u"test-user-uid"
-
-    def test_alert_rules_service_calls_add_user_with_expected_value(
+    def test_add_user_calls_alert_rules_service_add_user_with_expected_value(
         self, mock_alerts_service, mock_alert_rules_service
     ):
         alert_rules_client = AlertRulesClient(
             mock_alerts_service, mock_alert_rules_service
         )
-        alert_rules_client.add_user(self._rule_id, self._rule_id)
+        alert_rules_client.add_user(TEST_RULE_ID, TEST_RULE_ID)
         mock_alert_rules_service.add_user.assert_called_once_with(
-            self._rule_id, self._rule_id
+            TEST_RULE_ID, TEST_RULE_ID
         )
 
-    def test_alert_rules_service_raises_invalid_rule_type_error_when_adding_to_system_rule(
+    def test_add_user_raises_invalid_rule_type_error_when_adding_to_system_rule(
         self, mocker, mock_alerts_service_system_rule, mock_alert_rules_service
     ):
         def add(*args, **kwargs):
@@ -70,32 +68,75 @@ class TestAlertRulesClient(object):
             mock_alerts_service_system_rule, mock_alert_rules_service
         )
         with pytest.raises(Py42InvalidRuleOperationError) as err:
-            alert_rules_module.add_user(self._rule_id, self._rule_id)
+            alert_rules_module.add_user(TEST_RULE_ID, TEST_USER_ID)
 
-        assert (
-            "Only alert rules with a source of 'Alerting' can be targeted by this command."
-            in str(err.value)
+        expected_text = u"Unable to find or access Rule with ID '{}'. ".format(
+            TEST_RULE_ID
         )
+        expected_text += u"You might be trying to access a system rule."
+        assert expected_text in str(err.value)
 
-    def test_alert_rules_service_calls_remove_user_with_expected_value(
+    def test_remove_user_calls_alert_rules_service_remove_user_with_expected_value(
         self, mock_alerts_service, mock_alert_rules_service
     ):
         alert_rules_client = AlertRulesClient(
             mock_alerts_service, mock_alert_rules_service
         )
-        alert_rules_client.remove_user(self._rule_id, self._rule_id)
+        alert_rules_client.remove_user(TEST_RULE_ID, TEST_USER_ID)
         mock_alert_rules_service.remove_user.assert_called_once_with(
-            self._rule_id, self._rule_id
+            TEST_RULE_ID, TEST_USER_ID
         )
 
-    def test_alert_rules_service_calls_remove_all_users_with_expected_value(
+    def test_remove_user_raises_invalid_rule_type_error_when_adding_to_system_rule(
+        self, mocker, mock_alerts_service_system_rule, mock_alert_rules_service
+    ):
+        def add(*args, **kwargs):
+            base_err = mocker.MagicMock(spec=HTTPError)
+            base_err.response = mocker.MagicMock(spec=Response)
+            raise Py42NotFoundError(base_err)
+
+        mock_alert_rules_service.remove_user.side_effect = add
+        alert_rules_module = AlertRulesClient(
+            mock_alerts_service_system_rule, mock_alert_rules_service
+        )
+        with pytest.raises(Py42InvalidRuleOperationError) as err:
+            alert_rules_module.remove_user(TEST_RULE_ID, TEST_USER_ID)
+
+        expected_text = u"Unable to find or access Rule with ID '{}'. ".format(
+            TEST_RULE_ID
+        )
+        expected_text += u"You might be trying to access a system rule."
+        assert expected_text in str(err.value)
+
+    def test_remove_all_users_calls_alert_service_remove_all_users_with_expected_value(
         self, mock_alerts_service, mock_alert_rules_service
     ):
         alert_rules_client = AlertRulesClient(
             mock_alerts_service, mock_alert_rules_service
         )
-        alert_rules_client.remove_all_users(self._rule_id)
-        mock_alert_rules_service.remove_all_users.assert_called_once_with(self._rule_id)
+        alert_rules_client.remove_all_users(TEST_RULE_ID)
+        mock_alert_rules_service.remove_all_users.assert_called_once_with(TEST_RULE_ID)
+
+    def test_remove_all_users_raises_invalid_rule_type_error_when_adding_to_system_rule(
+        self, mocker, mock_alerts_service_system_rule, mock_alert_rules_service
+    ):
+        def add(*args, **kwargs):
+            base_err = mocker.MagicMock(spec=HTTPError)
+            base_err.response = mocker.MagicMock(spec=Response)
+            raise Py42NotFoundError(base_err)
+
+        mock_alert_rules_service.remove_all_users.side_effect = add
+        alert_rules_module = AlertRulesClient(
+            mock_alerts_service_system_rule, mock_alert_rules_service
+        )
+        with pytest.raises(Py42InvalidRuleOperationError) as err:
+            alert_rules_module.remove_all_users(TEST_RULE_ID)
+
+        expected_text = u"Unable to find or access Rule with ID '{}'. ".format(
+            TEST_RULE_ID
+        )
+        expected_text += u"You might be trying to access a system rule."
+        assert expected_text in str(err.value)
 
     def test_alert_rules_service_calls_get_all_with_expected_value(
         self, mock_alerts_service, mock_alert_rules_service